update test in ModelCrudTests to make assertions that would fail if something was broken
Prior to this change, the test was asserting that a Task and an ActionResult were not null, which of course,
will never occur, even if a breaking change were made to the code.

Also did a bit of code style cleanup because the squiggly lines trigger me and I can fix all occurrences with a single keystroke.